main/admin/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
pandas
openpyxl
reportlab
matplotlib
Werkzeug
transformers
torch
